Refactor validation to server side
Generating Hertz scaffolding code is not necessary, as the service
registry will handle the Generic Calls made to the RPC servers.

This commit shifts the routing logic to main.go, such that the
we no longer need the Hertz tool to generate codes.
This commit also adds a new middleware to the RPC server, which will
validate the incoming requests using the `thrift-gen-validator` plugin.

The HTTP server still checks if the JSON payload is valid, but now the
RPC server will check if the request parameters are valid.