Set up observability for the API Gateway
References for setting up and understanding observability:
https://github.com/cloudwego/hertz-examples/blob/main/opentelemetry/hertz/server/main.go
https://github.com/cloudwego/kitex-examples/blob/main/opentelemetry/README.md

https://opentelemetry.io/docs/concepts/
https://opentelemetry.io/docs/collector/configuration/

This commit makes use of the Kitex and Hertz OpenTelemetry libraries
to set up observability for the API Gateway. The observability
libraries are configured to send traces to the OpenTelemetry Collector
running in a Docker container. The OpenTelemetry Collector is configured
to send traces to the OpenTelemetry backend running in a Docker container.

The data can be viewed at http://localhost:3000/.