Restructure the project directory
This commit restructures the project directory based on the standards
defined in https://github.com/golang-standards/project-layout.

The new directory shifts the http-server and rpc-server to the
/examples directory.

The main usage of the project is to generate the code for the
API gateway and the service. The examples directory contains the
example of the generated code. The user can use the example to
build their own service.